megaminds-research/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Object Detection & Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0
pycocotools>=2.0.6

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
onnx-simplifier>=0.4.0

# Quantization (if using TensorRT)
# tensorrt>=8.6.0  # Uncomment if deploying to NVIDIA devices

# Training Utilities
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0
pyyaml>=6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# Evaluation & Metrics
scipy>=1.11.0
thop>=0.1.1  # For FLOPs calculation

# Research & Documentation
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
